requirements.txt
python-telegram-bot==22.3
openai==1.99.9
python-dotenv==1.1.1 